p29.26 refactor: thread ids.ProgramID everywhere
Typed-ID capstone: ProgramID is now a distinct type end-to-end. Purely types, no
behavior change; report goldens stay byte-identical.

Threaded (following the FundID template, p29.25):
- sqlc: Program.ID, ProgramsVersion.EntityID; every program-id query param/return
  (GetProgram, ProgramDescendants id + Row.ID, ProgramTreeRow.ID, InsertProgram
  return, InsertProgramVersionParams.ID_2, UpdateProgramParams.ID, ProgramSubtreeIDs
  param+return). Self-referential parent_id stays sql.NullInt64 at the scan boundary
  (Program/ProgramsVersion/*Row.ParentID) -- nullable FK.
- store: CreateProgram return, Create/UpdateProgramInput.ParentID, Update/Deactivate/
  Get/Descendants id params, ProgramPaths maps+return, checkDefaultProgram,
  insertProgramVersion, checkFundProgram; domain program-id fields typed on
  ProgramCell/FunctionalCellProgram/BudgetKeyCell/DescriptionPrefill (converted from
  the scanned NullInt64 at the boundary); nullable *ids.ProgramID on Create/Update
  Account/Fund/Split/BudgetSplit/ExpenseReportLine inputs, Register/Drill filters+rows,
  AccountEditorOption.DefaultProgram, ReportGrant.ProgramID, CurrentUser/UserSettings
  default program, GrantReportGroup + ProgramSubtree -- ids.Null on write,
  ids.Ptr[ids.ProgramID] on read.
- reports: reports.ProgramID alias already ids.ProgramID; Drill.ProgramID(s) +
  Encode/Decode (int64 at the URL string boundary), progCol.id/descendants, the
  program-statement/990 rollup drill cells.
- web: parse to typed at parseID seams; render-only model fields stay int64
  (int64(p.ID) at assignment, txnOption/programOption/regFilterOption/progRow ID);
  program-keyed store-boundary maps typed (programActivityTotals, depth, excluded,
  budget import progByName); grant-scope resolution typed (heldScope/wantScope/
  parseProgramScope/sameScope/clampProgramFilters/grantProgramScope, containsID made
  ~int64 generic); id->name render maps (programNameMap, admin progName) stay int64,
  cast at population.
- CLI/synth/fixture: ledgerimport BuildResult.ProgramIDs + rootProgramID;
  synth IDs.{General,Educacion,FoodPantry} + sp.prog/budget program/acctSpec.defProgram
  (alias-imported entids in build.go to dodge the `ids` IDs-param shadow); devseed
  firstProgram; fixture ProgramCell.Program + pkey.
- tests: program helper params/env fields typed; nullable-FK .Int64 comparisons wrap
  the typed side down (int64(prog)); AssertVersioned/itoa stay int64 -- callers pass
  int64(progID); budget_plans_test int64p -> progp (fund-typed sibling).

Boundary discipline (Rule 15): every nullable program FK stays sql.NullInt64 at the
scan boundary (splits/budget_splits/user_report_grants program_id, accounts
default_program_id, both parent_id self-FKs, history DiffValue.ID holder), typed only
at the store seam. SubsidiaryID/AccountID/SplitID/TransactionID unchanged.

+++ internal/reports/budget_reports.go
@@ -202,7 +202,7 @@ func actualDrill(key BudgetKey, bucketStart, bucketEnd, from, to string) *Drill 
 		tt = to
 	}
 	fund := key.Fund
-	prog := int64(key.Program)
+	prog := key.Program
 	return &Drill{
 		Scope:      int64(key.Subsidiary),
 		AccountIDs: []int64{int64(key.Account)},
